add "sbx help <cmd>"; parse command args
this adds `sbx help <cmd>` to get more info on what a
command does (for example, what flags it accepts).

this is a precursor to add a `duration` flag to `up`, as this sets us up
to add documented flags to the `up` command.

under the hood, this also adds the ability to parse additional args
for each command (right now, only `help` uses this).